requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
blis==1.0.1
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.8
databricks-sdk==0.57.0
datasets==3.6.0
Deprecated==1.2.15
dill==0.3.8
dnspython==2.7.0
docker==7.1.0
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.115.2
filelock==3.16.1
Flask==3.1.1
flatbuffers==24.12.23
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jwt==1.3.1
kiwisolver==1.4.8
kubernetes==31.0.0
langcodes==3.4.1
language_data==1.2.0
Mako==1.3.6
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.10.3
mdurl==0.1.2
mlflow==3.1.0
mlflow-skinny==3.1.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.4.2
numpy==2.0.2
oauthlib==3.2.2
onnx==1.18.0
onnxruntime==1.20.1
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pl_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.8.0/pl_core_news_sm-3.8.0-py3-none-any.whl#sha256=9b536db854b0cdb9132a74f68b392b8112ab6030f971bb03282462362a940dbb
posthog==3.7.4
preshed==3.0.9
propcache==0.3.2
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
sacremoses==0.1.1
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.3
srsly==2.4.8
starlette==0.40.0
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.66.5
transformers==4.48.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
waitress==3.0.2
wasabi==1.1.3
watchfiles==1.0.3
weasel==0.4.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0